--- tezos/baking/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "tezos_baking"
-version = "0.1.0"
-authors = ["Tomas Sedlak <tomas.sedlak@simplestaking.com>"]
-edition = "2018"
-
-[dependencies]
-failure = "0.1"
-# local dependencies
-crypto = { path = "../../crypto" }
-storage = { path = "../../storage" }
