--- tezos/baking/src/lib.rs
@@ -1,4 +1,0 @@
-// Copyright (c) SimpleStaking and Tezedge Contributors
-// SPDX-License-Identifier: MIT
-
-pub mod helpers;
